requirements.txt
tensorflow==2.12.1
pytest~=7.3.1
numpy~=1.24.3
pillow~=9.4.0
python-dotenv~=1.0.0
tensorboard~=2.12.3
colorama~=0.4.4
scikit-image~=0.20.0